requirements.txt
folium==0.11.0
geopandas==0.9.0
Pillow==9.0.0
streamlit==1.3.0
streamlit_folium==0.4.0
